Create separate component for TableCell.
